Share the side panel app path contract
