+++ packages/twenty-front/src/modules/object-record/record-index/components/RecordIndexPageHeader.tsx
@@ -7,14 +7,13 @@ import { useFilteredObjectMetadataItems } from '@/object-metadata/hooks/useFilte
 import { RecordIndexPageHeaderIcon } from '@/object-record/record-index/components/RecordIndexPageHeaderIcon';
 import { useRecordIndexContextOrThrow } from '@/object-record/record-index/contexts/RecordIndexContext';
 import { SidePanelToggleButton } from '@/side-panel/components/SidePanelToggleButton';
-import { useUpdateSidePanelPageInfo } from '@/side-panel/hooks/useUpdateSidePanelPageInfo';
+import { useSyncSidePanelPageTitle } from '@/side-panel/hooks/useUpdateSidePanelPageInfo';
 import { PageCardHeader } from '@/ui/layout/page/components/PageCardHeader';
 import { useWorkspaceSurface } from '@/ui/layout/hooks/useWorkspaceSurface';
 import { useAtomComponentStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomComponentStateValue';
 import { useAtomStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomStateValue';
 import { styled } from '@linaria/react';
 import { t } from '@lingui/core/macro';
-import { useEffect } from 'react';
 import { isDefined } from 'twenty-shared/utils';
 import { themeCssVariables } from 'twenty-ui/theme-constants';
 
@@ -36,7 +35,6 @@ const StyledSelectedRecordsCount = styled.div`
 
 export const RecordIndexPageHeader = () => {
   const workspaceSurface = useWorkspaceSurface();
-  const { updateSidePanelPageInfo } = useUpdateSidePanelPageInfo();
 
   const { findObjectMetadataItemByNamePlural } =
     useFilteredObjectMetadataItems();
@@ -54,11 +52,7 @@ export const RecordIndexPageHeader = () => {
 
   const label = objectMetadataItem?.labelPlural ?? objectNamePlural;
 
-  useEffect(() => {
-    if (workspaceSurface.type === 'side-panel') {
-      updateSidePanelPageInfo({ pageTitle: label });
-    }
-  }, [label, updateSidePanelPageInfo, workspaceSurface.type]);
+  useSyncSidePanelPageTitle(label);
 
   const pageHeaderTitle =
     contextStoreNumberOfSelectedRecords > 0 ? (
