+++ packages/twenty-front/src/modules/side-panel/hooks/useUpdateSidePanelPageInfo.ts
@@ -1,5 +1,6 @@
 import { sidePanelNavigationStackState } from '@/side-panel/states/sidePanelNavigationStackState';
-import { useCallback } from 'react';
+import { useWorkspaceSurface } from '@/ui/layout/hooks/useWorkspaceSurface';
+import { useCallback, useEffect } from 'react';
 import { type IconComponent } from 'twenty-ui/icon';
 import { useStore } from 'jotai';
 
@@ -43,4 +44,15 @@ export const useUpdateSidePanelPageInfo = () => {
   return {
     updateSidePanelPageInfo,
   };
+};
+
+export const useSyncSidePanelPageTitle = (pageTitle?: string) => {
+  const workspaceSurface = useWorkspaceSurface();
+  const { updateSidePanelPageInfo } = useUpdateSidePanelPageInfo();
+
+  useEffect(() => {
+    if (workspaceSurface.type === 'side-panel' && pageTitle !== undefined) {
+      updateSidePanelPageInfo({ pageTitle });
+    }
+  }, [pageTitle, updateSidePanelPageInfo, workspaceSurface.type]);
 };
