+++ packages/twenty-shared/src/types/index.ts
@@ -13,7 +13,8 @@ export { AggregateOperations } from './AggregateOperations';
 export type { AllowedFullNameSortSubField } from './AllowedFullNameSortSubField';
 export { ApiPath } from './ApiPath';
 export { AppBasePath } from './AppBasePath';
-export { AppPath } from './AppPath';
+export type { SidePanelAppPath } from './AppPath';
+export { AppPath, SIDE_PANEL_APP_PATHS } from './AppPath';
 export type { Arrayable } from './Arrayable';
 export type { ArraySortDirection } from './ArraySortDirection';
 export type { AssertUnreachable } from './AssertUnreachable.type';
